+++ src/main/java/ru/yandex/practicum/filmorate/controller/exception/ItemNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.controller.exception;
 
 public class ItemNotFoundException extends RuntimeException {
-    public ItemNotFoundException(int itemId, String itemName) {
+    public ItemNotFoundException(long itemId, String itemName) {
         super(String.format("%s с id = %d не существует", itemName, itemId));
     }
 }
